デバッグ用データを削除
